feat: Change int to int8_t to handle single byte instead of 4-bytes ints

+++ c/chunk_type.h
@@ -1,18 +1,19 @@
+#include <stdint.h>
 #ifndef _CHUNKTYPEH_
 #define _CHUNKTYPEH_
 
 typedef struct {
-  int ancillary;
-  int priv;
-  int reserved;
-  int safe;
+  int8_t ancillary;
+  int8_t priv;
+  int8_t reserved;
+  int8_t safe;
 } ChunkType;
 
-int* to_bytes(ChunkType* type);
-int is_valid(ChunkType* type);
-int is_critical(ChunkType* type);
-int is_reserved_bit_valid(ChunkType* type);
-int is_safe_to_copy(ChunkType* type);
-ChunkType from_bytes(int bytes[]);
+int8_t* to_bytes(ChunkType* type);
+int8_t is_valid(ChunkType* type);
+int8_t is_critical(ChunkType* type);
+int8_t is_reserved_bit_valid(ChunkType* type);
+int8_t is_safe_to_copy(ChunkType* type);
+ChunkType from_bytes(int8_t bytes[]);
 
 #endif
